refactor: remove gender and add appearance in test
