+++ test/api/character/test_get_character_by_id.py
