+++ test/api/character/test_patch_character.py
